Add Security for friends end points

+++ backend/src/user-relations/user-relations.service.ts
@@ -1,6 +1,10 @@
 import { Injectable } from '@nestjs/common'
 import { PrismaService } from 'src/prisma/prisma.service'
-import { EUserRelationType, Prisma } from '@prisma/client'
+import { EUserRelationType } from '@prisma/client'
+import {
+  BadRequestException,
+  UnauthorizedException,
+} from '@nestjs/common/exceptions'
 
 @Injectable()
 export class UserRelationsService {
@@ -9,32 +13,42 @@ export class UserRelationsService {
   //  MUTATION
   //**************************************************//
 
-  async create(data: Prisma.UserRelationUncheckedCreateInput) {
-    await this.prisma.userRelation.create({ data })
-    const tmp = data.userOwnerId
-    data.userOwnerId = data.userTargetId
-    data.userTargetId = tmp
-    return await this.prisma.userRelation.create({ data })
-  }
+  async createRequestFriend(userOwnerId: string, userTargetId: string) {
+    const relationOT = await this.findRelation(userOwnerId, userTargetId)
+    const relationTO = await this.findRelation(userTargetId, userOwnerId)
 
-  async update(
-    userOwnerId: string,
-    userTargetId: string,
-    data: Prisma.UserRelationUncheckedUpdateInput,
-  ) {
-    return await this.prisma.userRelation.update({
-      where: {
-        userOwnerId_userTargetId: {
-          userOwnerId,
-          userTargetId,
-        },
+    if (relationTO === EUserRelationType.PendingAccept)
+      return await this.acceptPendingRelation(userOwnerId, userTargetId)
+    if (relationOT !== undefined || relationTO !== undefined)
+      throw new UnauthorizedException()
+
+    await this.prisma.userRelation.create({
+      data: {
+        userOwnerId: userTargetId,
+        userTargetId: userOwnerId,
+        type: EUserRelationType.WaitingAccept,
       },
-      data,
+    })
+    return await this.prisma.userRelation.create({
+      data: {
+        userOwnerId,
+        userTargetId,
+        type: EUserRelationType.PendingAccept,
+      },
     })
   }
 
   // only works if the invite is in pending mode, do not work if user is blocked
   async acceptPendingRelation(userOwnerId: string, userTargetId: string) {
+    const relationOT = await this.findRelation(userOwnerId, userTargetId)
+    const relationTO = await this.findRelation(userTargetId, userOwnerId)
+
+    if (
+      relationOT !== EUserRelationType.WaitingAccept &&
+      relationTO !== EUserRelationType.PendingAccept
+    )
+      throw new UnauthorizedException()
+
     await this.prisma.userRelation.update({
       where: {
         userOwnerId_userTargetId: {
@@ -59,22 +73,40 @@ export class UserRelationsService {
     })
   }
 
-  async RefusePendingRelation(userOwnerId: string, userTargetId: string) {
-    return await this.deleteBoth(userOwnerId, userTargetId)
-  }
+  async refusePendingRelation(userOwnerId: string, userTargetId: string) {
+    const relationOT = await this.findRelation(userOwnerId, userTargetId)
+    const relationTO = await this.findRelation(userTargetId, userOwnerId)
 
-  async block(userOwnerId: string, userTargetId: string) {
-    await this.prisma.userRelation.update({
+    if (
+      relationOT !== EUserRelationType.WaitingAccept &&
+      relationTO !== EUserRelationType.PendingAccept
+    )
+      throw new UnauthorizedException()
+
+    return await this.prisma.userRelation.delete({
       where: {
-        userOwnerId_userTargetId: {
-          userTargetId: userOwnerId,
-          userOwnerId: userTargetId,
-        },
-      },
-      data: {
-        type: EUserRelationType.Blocked,
+        userOwnerId_userTargetId: { userOwnerId, userTargetId },
       },
     })
+  }
+
+  async blockRelation(userOwnerId: string, userTargetId: string) {
+    const relationOT = await this.findRelation(userOwnerId, userTargetId)
+    const relationTO = await this.findRelation(userTargetId, userOwnerId)
+
+    if (relationOT === EUserRelationType.Blocked)
+      throw new BadRequestException()
+
+    if (relationTO !== EUserRelationType.Blocked) {
+      await this.prisma.userRelation.delete({
+        where: {
+          userOwnerId_userTargetId: {
+            userTargetId: userOwnerId,
+            userOwnerId: userTargetId,
+          },
+        },
+      })
+    }
     return await this.prisma.userRelation.update({
       where: {
         userOwnerId_userTargetId: {
@@ -88,31 +120,45 @@ export class UserRelationsService {
     })
   }
 
-  async deleteOne(userOwnerId: string, userTargetId: string) {
+  async unblockRelation(userOwnerId: string, userTargetId: string) {
+    const relationOT = await this.findRelation(userOwnerId, userTargetId)
+
+    if (relationOT !== EUserRelationType.Blocked)
+      throw new BadRequestException()
+
     return await this.prisma.userRelation.delete({
       where: {
         userOwnerId_userTargetId: {
-          userOwnerId,
           userTargetId,
+          userOwnerId,
         },
       },
     })
   }
 
-  async deleteBoth(userOwnerId: string, userTargetId: string) {
+  async removeFriend(userOwnerId: string, userTargetId: string) {
+    const relationOT = await this.findRelation(userOwnerId, userTargetId)
+    const relationTO = await this.findRelation(userTargetId, userOwnerId)
+
+    if (
+      relationOT !== EUserRelationType.Friend ||
+      relationTO !== EUserRelationType.Friend
+    )
+      throw new BadRequestException()
+
     await this.prisma.userRelation.delete({
       where: {
         userOwnerId_userTargetId: {
-          userOwnerId,
-          userTargetId,
+          userOwnerId: userTargetId,
+          userTargetId: userOwnerId,
         },
       },
     })
     return await this.prisma.userRelation.delete({
       where: {
         userOwnerId_userTargetId: {
-          userOwnerId: userTargetId,
-          userTargetId: userOwnerId,
+          userOwnerId,
+          userTargetId,
         },
       },
     })
@@ -122,6 +168,21 @@ export class UserRelationsService {
   //  QUERY
   //**************************************************//
 
+  private async findRelation(
+    userAId: string,
+    userBId: string,
+  ): Promise<EUserRelationType | undefined> {
+    const relation = await this.prisma.userRelation.findUnique({
+      where: {
+        userOwnerId_userTargetId: {
+          userOwnerId: userAId,
+          userTargetId: userBId,
+        },
+      },
+    })
+    return relation?.type
+  }
+
   async findAll() {
     return await this.prisma.userRelation.findMany({})
   }
@@ -130,7 +191,7 @@ export class UserRelationsService {
     return await this.prisma.userRelation.findUnique({
       where: {
         userOwnerId_userTargetId: {
-          userOwnerId,
+          userOwnerId: userOwnerId,
           userTargetId,
         },
       },
